backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
gunicorn==21.2.0